feat: TouristSpot과 recruit 연결

+++ src/main/java/com/multi/ouigo/domain/recruit/dto/req/CreateRecruitReqDto.java
@@ -37,7 +37,7 @@ public class CreateRecruitReqDto {
 
     private List<AgeCode> ageCodes;
 
-    private int touristSpotId;
+    private Long touristSpotId;
 
 
 }
